# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Redoubt API
  description: |
    Redoubt is a secure, invite-only communication platform API.

    All error responses conform to RFC 9457 Problem Details for HTTP APIs.
    All timestamps are in ISO 8601 UTC format.
    All IDs are UUID v4.
  version: 0.1.0
  contact:
    name: Redoubt
  license:
    name: MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "https://redoubt.app/problems/validation-error"
        title:
          type: string
          description: Short, human-readable summary
          example: "Validation Failed"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation
          example: "The request body contains invalid fields"
        instance:
          type: string
          description: URI reference for this occurrence
          example: "/api/v1/auth/register"
        request_id:
          type: string
          description: X-Request-ID for tracing
          example: "abc-123-def"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: "username"
        message:
          type: string
          example: "must be between 3 and 32 characters"

    # Auth schemas
...
    ReactionGroup:
      type: object
      properties:
        emoji:
          type: string
          example: "👍"
        count:
          type: integer
          example: 3
        users:
          type: array
          items:
            type: string
          description: Usernames of users who reacted
        has_reacted:
          type: boolean
          description: Whether the current user has this reaction
    ReactionsListResponse:
      type: object
      properties:
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/ReactionGroup'
    AddReactionRequest:
      type: object
      required:
        - emoji
      properties:
        emoji:
          type: string
          minLength: 1
          maxLength: 32
          example: "👍"
    ToggleReactionResponse:
      type: object
      properties:
        added:
          type: boolean
          description: Whether the reaction was added (true) or removed (false)

    # Emoji schemas
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests allowed in window
        RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in current window
        RateLimit-Reset:
          schema:
            type: integer
          description: Seconds until the window resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
paths:
...
  /messages/{id}/reactions:
    get:
      tags: [Reactions]
      summary: Get message reactions
      description: Get all reactions for a message, grouped by emoji.
      operationId: getMessageReactions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
          description: Message ID
      responses:
        '200':
          description: Message reactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Reactions]
      summary: Add reaction
      description: Add a reaction to a message. Emoji must be from the curated set. Rate limited to 20 reactions per minute.
      operationId: addReaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
          description: Message ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddReactionRequest'
      responses:
        '201':
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /messages/{id}/reactions/toggle:
    post:
      tags: [Reactions]
      summary: Toggle reaction
      description: Add a reaction if it doesn't exist, or remove it if it does.
      operationId: toggleReaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
          description: Message ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddReactionRequest'
      responses:
        '200':
          description: Reaction toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToggleReactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /messages/{id}/reactions/{emoji}:
    delete:
      tags: [Reactions]
      summary: Remove reaction
      description: Remove your reaction from a message.
      operationId: removeReaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
          description: Message ID
        - name: emoji
          in: path
          required: true
          schema:
            type: string
          description: The emoji to remove (URL-encoded)
      responses:
        '204':
          description: Reaction removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Emoji endpoints
...